Added Admin route and page
